Refactor x/staking according to ADR 031
